pyproject.toml
[tool.poetry]
name = "learnprotobuf"
version = "0.1.0"
description = ""
authors = ["ntbloom <ntbloom@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
protobuf = "^3.17.3"
pyserial = "^3.5"
black = "^21.7b0"
hypothesis = "^6.14.6"
pytest = "^6.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
